% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropart.R
\name{as.ProbaVector.pattern_matrix_individuals}
\alias{as.ProbaVector.pattern_matrix_individuals}
\title{Probabilty vector of a matrix of individuals}
\usage{
\method{as.ProbaVector}{pattern_matrix_individuals}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{pattern_matrix_individuals}}.}

\item{...}{Further arguments. Unsused.}
}
\value{
A \code{\link{ProbaVector}} object, which is a named integer vector.
}
\description{
Gets an \code{\link{ProbaVector}} object from a \code{\link{pattern_matrix_individuals}} object to measure diversity with the entropart package.
}
