% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{autoplot.community_model}
\alias{autoplot.community_model}
\title{autoplot Community Models}
\usage{
\method{autoplot}{community_model}(object, ...)
}
\arguments{
\item{object}{The model to plot.}

\item{...}{Extra arguments passed to the plot method.}
}
\description{
S3 method to force the use of the R6 method.
}
\details{
This is a helper function to always use the R6 method rather than the S3 method:
\code{autoplot(MyModel)} is translated to \code{MyModel$autoplot}.
}
\examples{
myModel <- cm_drift$new(pattern_matrix_individuals(S = 10))
autoplot(myModel)
}
