% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_Models.R
\docType{class}
\name{cg_drift}
\alias{cg_drift}
\title{Community Drift, grid version}
\description{
A \link{community_gridmodel} where each cell contains an individual.
Marks are species.
At each generation, each individual is replaced by one of its neighbors.
}
\section{Super classes}{
\code{\link[SimComm:community_model]{SimComm::community_model}} -> \code{\link[SimComm:community_gridmodel]{SimComm::community_gridmodel}} -> \code{cg_drift}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cg_drift-clone}{\code{cg_drift$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="along_time"><a href='../../SimComm/html/community_model.html#method-community_model-along_time'><code>SimComm::community_model$along_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="autoplot"><a href='../../SimComm/html/community_model.html#method-community_model-autoplot'><code>SimComm::community_model$autoplot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="run"><a href='../../SimComm/html/community_model.html#method-community_model-run'><code>SimComm::community_model$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="saved_pattern"><a href='../../SimComm/html/community_model.html#method-community_model-saved_pattern'><code>SimComm::community_model$saved_pattern()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_gridmodel" data-id="initialize"><a href='../../SimComm/html/community_gridmodel.html#method-community_gridmodel-initialize'><code>SimComm::community_gridmodel$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_gridmodel" data-id="neighbors"><a href='../../SimComm/html/community_gridmodel.html#method-community_gridmodel-neighbors'><code>SimComm::community_gridmodel$neighbors()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_gridmodel" data-id="plot"><a href='../../SimComm/html/community_gridmodel.html#method-community_gridmodel-plot'><code>SimComm::community_gridmodel$plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cg_drift-clone"></a>}}
\if{latex}{\out{\hypertarget{method-cg_drift-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cg_drift$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
