% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{community_gridmodel}
\alias{community_gridmodel}
\title{Community Grid Model Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
community_gridmodel
}
\value{
An \code{\link{R6Class}}.
}
\description{
A \code{\link{community_model}} whose pattern is a regular, rectangular grid of points.
}
\section{Fields}{

\describe{
\item{\code{tess}}{A tesselization of the window containing the points, used to plot the model.}
}}

\section{Methods}{

\describe{
  \item{\code{initialize(pattern = pattern_matrix(), evolve = evolve_none, timeline = timeline_regular())}}{Initialization.}
  \item{\code{plot(...)}}{Default plot method: plots the pattern.}
  \item{\code{run(animate = FALSE, sleep = 0)}}{Run the model.}
}
}

\keyword{datasets}
