% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{community_matrixmodel}
\alias{community_matrixmodel}
\title{Community Matrix Model Class}
\description{
Community Matrix Model Class

Community Matrix Model Class
}
\details{
A \link{community_model} whose pattern is a regular, rectangular grid of points.
}
\section{Super class}{
\code{\link[SimComm:community_model]{SimComm::community_model}} -> \code{community_matrixmodel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{neighborhood}}{A character string defining what is the neighborhood of a cell:
"von Neumann 1" or "4" for the closest four neighbors (North, West, South, East);
"Moore 1" or "8" for all adjacent cells (the first four and North-West, etc.);
"Moore 2" or "24" for two rings of neighbors.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-community_matrixmodel-new}{\code{community_matrixmodel$new()}}
\item \href{#method-community_matrixmodel-prepare_buffer}{\code{community_matrixmodel$prepare_buffer()}}
\item \href{#method-community_matrixmodel-neighbors}{\code{community_matrixmodel$neighbors()}}
\item \href{#method-community_matrixmodel-plot}{\code{community_matrixmodel$plot()}}
\item \href{#method-community_matrixmodel-autoplot}{\code{community_matrixmodel$autoplot()}}
\item \href{#method-community_matrixmodel-clone}{\code{community_matrixmodel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="along_time"><a href='../../SimComm/html/community_model.html#method-community_model-along_time'><code>SimComm::community_model$along_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="run"><a href='../../SimComm/html/community_model.html#method-community_model-run'><code>SimComm::community_model$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="saved_pattern"><a href='../../SimComm/html/community_model.html#method-community_model-saved_pattern'><code>SimComm::community_model$saved_pattern()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_matrixmodel-new"></a>}}
\if{latex}{\out{\hypertarget{method-community_matrixmodel-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_matrixmodel$new(
  pattern = pattern_matrix_individuals(),
  timeline = 0,
  type = "Species",
  neighborhood = "von Neumann 1"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pattern}}{The pattern which describes the location of agents.}

\item{\code{pattern}}{The pattern which describes the location of agents.}

\item{\code{timeline}}{A numeric vector that contains the points of time of interest.}

\item{\code{type}}{The type of individuals. Informational only.}

\item{\code{neighborhood}}{A character string defining what is the neighborhood of a cell:
"von Neumann 1" or "4" for the closest four neighbors (North, West, South, East);
"Moore 1" or "8" for all adjacent cells (the first four and North-West, etc.);
"Moore 2" or "24" for two rings of neighbors.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_matrixmodel-prepare_buffer"></a>}}
\if{latex}{\out{\hypertarget{method-community_matrixmodel-prepare_buffer}{}}}
\subsection{Method \code{prepare_buffer()}}{
Prepare buffered_pattern at each step of evolution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_matrixmodel$prepare_buffer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_matrixmodel-neighbors"></a>}}
\if{latex}{\out{\hypertarget{method-community_matrixmodel-neighbors}{}}}
\subsection{Method \code{neighbors()}}{
Return the vector of neighbors of a cell, defined by it row and column
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_matrixmodel$neighbors(row, col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{row}}{The row of the focal cell.}

\item{\code{col}}{The column of the focal cell.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_matrixmodel-plot"></a>}}
\if{latex}{\out{\hypertarget{method-community_matrixmodel-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the pattern.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_matrixmodel$plot(time = NULL, sleep = 0, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The point of the timeline considered.
Its value should be in \code{timeline}.}

\item{\code{sleep}}{The time (in ms) to sleep between each step of the evolution of the model.}

\item{\code{...}}{Extra arguments to be passed to methods.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_matrixmodel-autoplot"></a>}}
\if{latex}{\out{\hypertarget{method-community_matrixmodel-autoplot}{}}}
\subsection{Method \code{autoplot()}}{
Produces a ggplot of the pattern.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_matrixmodel$autoplot(time = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The point of the timeline considered.
Its value should be in \code{timeline}.}

\item{\code{...}}{Extra arguments to be passed to methods.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_matrixmodel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-community_matrixmodel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_matrixmodel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
