% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco2.R
\docType{class}
\name{community_param2}
\alias{community_param2}
\title{Community Attributes Class}
\description{
The class generate a community of a large number of species.
\subsection{this part isn't complete yet}{
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-community_param2-new}{\code{community_param2$new()}}
\item \href{#method-community_param2-set_values}{\code{community_param2$set_values()}}
\item \href{#method-community_param2-show_values}{\code{community_param2$show_values()}}
\item \href{#method-community_param2-draw_matrix}{\code{community_param2$draw_matrix()}}
\item \href{#method-community_param2-draw_wmppp}{\code{community_param2$draw_wmppp()}}
\item \href{#method-community_param2-clone}{\code{community_param2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_param2-new"></a>}}
\if{latex}{\out{\hypertarget{method-community_param2-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_param2$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_param2-set_values"></a>}}
\if{latex}{\out{\hypertarget{method-community_param2-set_values}{}}}
\subsection{Method \code{set_values()}}{
Set values of the community generation parameters (set to default if no arguments)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_param2$set_values(
  nx = 20,
  ny = nx,
  S = 2,
  Distribution = "lnorm",
  sd = 1,
  prob = 0.1,
  alpha = 40
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nx}}{The size of horizontal axis}

\item{\code{ny}}{The size of vertical axis}

\item{\code{S}}{The number of species.

For the community drawing with entropart}

\item{\code{Distribution}}{The distribution of species frequencies. May be "lnorm" (log-normal), "lseries" (log-series), "geom" (geometric) or "bstick" (broken stick).}

\item{\code{sd}}{The simulated distribution standard deviation. For the log-normal distribution, this is the standard deviation on the log scale.}

\item{\code{prob}}{The proportion of ressources taken by successive species in the geometric model.}

\item{\code{alpha}}{Fisher's alpha in the log-series model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_param2-show_values"></a>}}
\if{latex}{\out{\hypertarget{method-community_param2-show_values}{}}}
\subsection{Method \code{show_values()}}{
Show the values of the community generation parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_param2$show_values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_param2-draw_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-community_param2-draw_matrix}{}}}
\subsection{Method \code{draw_matrix()}}{
Draw a community with no adjectives
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_param2$draw_matrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_param2-draw_wmppp"></a>}}
\if{latex}{\out{\hypertarget{method-community_param2-draw_wmppp}{}}}
\subsection{Method \code{draw_wmppp()}}{
Draw a wmppp community with no adjective
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_param2$draw_wmppp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_param2-clone"></a>}}
\if{latex}{\out{\hypertarget{method-community_param2-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_param2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
