% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{community_spcmodel}
\alias{community_spcmodel}
\title{Community Point Pattern Class}
\value{
An \code{\link{R6Class}}.
}
\description{
A \code{\link{community_model}} whose pattern is a \code{\link[dbmss]{wmppp}} object..
}
\section{Public Methods}{

\describe{
  \item{\code{initialize(pattern = NULL, timeline = 0, type = "Species")}}{Initialization.}
  \item{\code{plot(time=NULL, sleep=animation::ani.options("interval"), ...)}}{Default plot method: plots the pattern.}
  \item{\code{autoplot(time=NULL, ...)}}{Makes a \code{\link{ggplot}} of the pattern.}
  \item{\code{run(animate = FALSE, sleep = animation::ani.options("interval"), save = FALSE, more_time = NULL)}}{Run the model.}
  \item{\code{saved_pattern(time)}}{Returns the pattern at the chosen time.}
  \item{\code{along_time(FUN, ...)}}{Applies the function FUN to the saved patterns along time and returns a dataframe with columns \code{x} for time and \code{y} for the results of FUN. FUN must return a single value.}
}
}

\examples{
myModel <- community_matrixmodel$new()
myModel$autoplot()
}
\section{Super class}{
\code{\link[SimComm:community_model]{SimComm::community_model}} -> \code{community_spcmodel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{community_spcmodel$new()}}
\item \href{#method-neighbors_n}{\code{community_spcmodel$neighbors_n()}}
\item \href{#method-neighbors_r}{\code{community_spcmodel$neighbors_r()}}
\item \href{#method-neighbor_types_r}{\code{community_spcmodel$neighbor_types_r()}}
\item \href{#method-plot}{\code{community_spcmodel$plot()}}
\item \href{#method-clone}{\code{community_spcmodel$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="along_time">}\href{../../SimComm/html/community_model.html#method-along_time}{\code{SimComm::community_model$along_time()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="autoplot">}\href{../../SimComm/html/community_model.html#method-autoplot}{\code{SimComm::community_model$autoplot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="run">}\href{../../SimComm/html/community_model.html#method-run}{\code{SimComm::community_model$run()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="saved_pattern">}\href{../../SimComm/html/community_model.html#method-saved_pattern}{\code{SimComm::community_model$saved_pattern()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$new(
  pattern = SpatDiv::rSpCommunity(n = 1, size = 100, CheckArguments = FALSE),
  timeline = 0,
  type = "Species"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-neighbors_n"></a>}}
\if{latex}{\out{\hypertarget{method-neighbors_n}{}}}
\subsection{Method \code{neighbors_n()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$neighbors_n(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-neighbors_r"></a>}}
\if{latex}{\out{\hypertarget{method-neighbors_r}{}}}
\subsection{Method \code{neighbors_r()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$neighbors_r(r, keep_distances = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-neighbor_types_r"></a>}}
\if{latex}{\out{\hypertarget{method-neighbor_types_r}{}}}
\subsection{Method \code{neighbor_types_r()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$neighbor_types_r(r)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$plot(time = NULL, sleep = 0, which.marks = "PointType", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
