% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_Models.R
\docType{class}
\name{cp_drift}
\alias{cp_drift}
\title{Community Drift}
\value{
An \code{\link{R6Class}}.
}
\description{
A \code{\link{community_spcmodel}}.
At each generation, each individual is replaced by one of its neghbors.
}
\section{Public Methods}{

\describe{
  \item{\code{initialize(pattern = NULL, timeline = 0, type = "Species")}}{Initialization.}
  \item{\code{plot(time=NULL, sleep=animation::ani.options("interval"), ...)}}{Default plot method: plots the pattern.}
  \item{\code{autoplot(time=NULL, ...)}}{Makes a \code{\link{ggplot}} of the pattern.}
  \item{\code{run(animate = FALSE, sleep = animation::ani.options("interval"), save = FALSE, more_time = NULL)}}{Run the model.}
  \item{\code{saved_pattern(time)}}{Returns the pattern at the chosen time.}
  \item{\code{along_time(FUN, ...)}}{Applies the function FUN to the saved patterns along time and returns a dataframe with columns \code{x} for time and \code{y} for the results of FUN. FUN must return a single value.}
}
}

\examples{
myModel <- community_matrixmodel$new()
myModel$autoplot()
}
\section{Super classes}{
\code{\link[SimComm:community_model]{SimComm::community_model}} -> \code{\link[SimComm:community_spcmodel]{SimComm::community_spcmodel}} -> \code{cp_drift}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n_neighbors}}{The number of nearest neighbors to take into account.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cp_drift-new}{\code{cp_drift$new()}}
\item \href{#method-cp_drift-clone}{\code{cp_drift$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="along_time"><a href='../../SimComm/html/community_model.html#method-community_model-along_time'><code>SimComm::community_model$along_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="autoplot"><a href='../../SimComm/html/community_model.html#method-community_model-autoplot'><code>SimComm::community_model$autoplot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="run"><a href='../../SimComm/html/community_model.html#method-community_model-run'><code>SimComm::community_model$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="saved_pattern"><a href='../../SimComm/html/community_model.html#method-community_model-saved_pattern'><code>SimComm::community_model$saved_pattern()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_spcmodel" data-id="neighbor_types_r"><a href='../../SimComm/html/community_spcmodel.html#method-community_spcmodel-neighbor_types_r'><code>SimComm::community_spcmodel$neighbor_types_r()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_spcmodel" data-id="neighbors_n"><a href='../../SimComm/html/community_spcmodel.html#method-community_spcmodel-neighbors_n'><code>SimComm::community_spcmodel$neighbors_n()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_spcmodel" data-id="neighbors_r"><a href='../../SimComm/html/community_spcmodel.html#method-community_spcmodel-neighbors_r'><code>SimComm::community_spcmodel$neighbors_r()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_spcmodel" data-id="plot"><a href='../../SimComm/html/community_spcmodel.html#method-community_spcmodel-plot'><code>SimComm::community_spcmodel$plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cp_drift-new"></a>}}
\if{latex}{\out{\hypertarget{method-cp_drift-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cp_drift$new(
  pattern = SpatDiv::rSpCommunity(n = 1, size = 100, CheckArguments = FALSE),
  timeline = 0,
  type = "Species",
  n_neighbors = 6
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cp_drift-clone"></a>}}
\if{latex}{\out{\hypertarget{method-cp_drift-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cp_drift$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
