% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropart.R
\name{entropart}
\alias{entropart}
\alias{as.AbdVector.pattern_matrix_individuals}
\alias{as.ProbaVector.pattern_matrix_individuals}
\alias{Tsallis.pattern_matrix_individuals}
\alias{Diversity.pattern_matrix_individuals}
\alias{Richness.pattern_matrix_individuals}
\alias{Shannon.pattern_matrix_individuals}
\alias{Simpson.pattern_matrix_individuals}
\title{entropart Methods}
\usage{
\method{as.AbdVector}{pattern_matrix_individuals}(x, ...)

\method{as.ProbaVector}{pattern_matrix_individuals}(x, ...)

\method{Tsallis}{pattern_matrix_individuals}(NorP, q = 1, Correction = "Best", ..., CheckArguments = TRUE)

\method{Diversity}{pattern_matrix_individuals}(NorP, q = 1, Correction = "Best", ..., CheckArguments = TRUE)

\method{Richness}{pattern_matrix_individuals}(
  NorP,
  Correction = "Chao1",
  Alpha = 0.05,
  JackOver = FALSE,
  ...,
  CheckArguments = TRUE
)

\method{Shannon}{pattern_matrix_individuals}(NorP, Correction = "Best", ..., CheckArguments = TRUE)

\method{Simpson}{pattern_matrix_individuals}(NorP, Correction = "Lande", ..., CheckArguments = TRUE)
}
\arguments{
\item{x}{An object of class \code{\link{pattern_matrix_individuals}}.}

\item{...}{Further arguments. Unsused.}

\item{NorP}{An object of class "wmppp" (\code{\link{pattern_matrix_individuals}}).}

\item{q}{A number: the order of entropy. Some corrections allow only a positive number. Default is 1 for Shannon entropy.}

\item{Correction}{A string containing one of the possible corrections: \code{"None"} (no correction), and \code{"Best"} are always valid. See the generic function help for other posiibilities.}

\item{CheckArguments}{If \code{TRUE} (default), the function arguments are verified. Should be set to \code{FALSE} to save time in simulations for example, when the arguments have been checked elsewhere.}

\item{Alpha}{The risk level, 5\% by default, used to optimize the jackknife order.}

\item{JackOver}{If \code{TRUE}, retain the jackknife order immediately superior to the optimal one, usually resulting in the overestimation of the number of species. Default is \code{FALSE}.}
}
\description{
The entropart package provides generic methods to measure diversity.
S3 methods to aplly them to simulated communities are available here.
\code{\link{AbdVector}} returns an abundance vector and \code{\link{ProbaVector}} returns a probability vector.
\code{\link{Tsallis}} returns Tsallis's entropy of a community, \code{\link{Diversity}} its diversity.
\code{\link{Richness}}, \code{\link{Shannon}}  and \code{\link{Simpson}} return particular indices of diversity.
}
\examples{
# A community matrix drift model
myModel <- cm_drift$new(pattern_matrix_individuals(S=10))
myModel$autoplot()
as.AbdVector(myModel$pattern)
as.ProbaVector(myModel$pattern)
Tsallis(myModel$pattern)
Diversity(myModel$pattern)
Richness(myModel$pattern)
Shannon(myModel$pattern)
Simpson(myModel$pattern)
}
