% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evolve.R
\name{evolve}
\alias{evolve}
\alias{evolve_none}
\alias{evolve_grid_byneighbors}
\alias{evolve_matrix_byneighbors}
\title{Model Evolution}
\usage{
evolve_none(deltaT)

evolve_grid_byneighbors(deltaT, type = "von Neumann")

evolve_matrix_byneighbors(deltaT, type = "von Neumann")
}
\arguments{
\item{deltaT}{The time passed since the last step of the evolution of the model.}

\item{type}{Definition of the neighborhhod. "von Neumann" or "4" means the upper, lower, left and right neighbors are considered;}
}
\value{
\code{evolve} methods should modifiy the model's pattern directly and return \code{NULL}.
}
\description{
At each time of the \code{timeline}, the \code{evolve} method of a \code{\link{community_model}} is run.
It is a function that modifies the \code{pattern} of the model.
Several \code{evolve_xxx} functions are available here: they are to be used as \code{evolve} methods.
}
\details{
Available functions:
\describe{
  \item{\code{evolve_none}}{ does nothing.}
  \item{\code{evolve_grid_byneighbors}}{Changes the type of each point of the grif by the type of one of its neighbors.}
}
}
