% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{graphs}
\alias{graphs}
\title{Graphs}
\description{
Graphs describe the structure of the community.
}
\details{
Several graphs are available by default.

Available graphs are:
\itemize{
\item \code{timed_abundance}
\item \code{timed_relative_abundance}
\item \code{overall_abundance_distribution}
\item \code{overall_abundance_rank}
\item \code{relative_abundance_rank}
}
}
