% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Communities.R
\name{local_pc}
\alias{local_pc}
\title{Local}
\description{
Local

Local
}
\section{Super class}{
\code{SimComm::community_param} -> \code{local_pc}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{death_rate}}{The mortality rate of an individual.
Default is \code{0.1}}

\item{\code{birth_rate}}{The reproduction rate of an individual.
Default is \code{0.2}}

\item{\code{speciation_rate}}{The speciation rate of an individual
Default is \code{0.00001}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-local_pc-new}{\code{local_pc$new()}}
\item \href{#method-local_pc-make_local}{\code{local_pc$make_local()}}
\item \href{#method-local_pc-clone}{\code{local_pc$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param" data-id="draw_matrix"><a href='../../SimComm/html/community_param.html#method-community_param-draw_matrix'><code>SimComm::community_param$draw_matrix()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param" data-id="draw_wmppp"><a href='../../SimComm/html/community_param.html#method-community_param-draw_wmppp'><code>SimComm::community_param$draw_wmppp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param" data-id="set_values"><a href='../../SimComm/html/community_param.html#method-community_param-set_values'><code>SimComm::community_param$set_values()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-local_pc-new"></a>}}
\if{latex}{\out{\hypertarget{method-local_pc-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link[R6:R6Class]{R6} class.
TODO : add more ecological rates
do a documentation on it, its a bit confusing right now
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_pc$new(
  fashion = "matrix",
  death_rate = self$death_rate,
  speciation_rate = self$speciation_rate,
  draw = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-local_pc-make_local"></a>}}
\if{latex}{\out{\hypertarget{method-local_pc-make_local}{}}}
\subsection{Method \code{make_local()}}{
Local community default community drawing - add death/birth_rate in the matrix
TODO : Verify if defaulting values in argument is right
I could do an option and make a set_value ?
strengthen the code, verify if the rates are on same level IN the matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_pc$make_local(
  nx = 10,
  ny = nx,
  S = 2,
  Distribution = "lnorm",
  sd = 1,
  prob = 0.1,
  theta = 40,
  death_rate = 0.1,
  birth_rate = 0.2,
  draw = FALSE,
  fashion = "matrix"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-local_pc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-local_pc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_pc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
