% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco2.R
\docType{class}
\name{local_pc2}
\alias{local_pc2}
\title{Local community}
\description{
The class generate a local community of a large number of species.
}
\section{Super class}{
\code{SimComm::community_param2} -> \code{local_pc2}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{death_rate}}{The mortality rate of an individual.
Default is \code{0.1}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-local_pc2-new}{\code{local_pc2$new()}}
\item \href{#method-local_pc2-make_local}{\code{local_pc2$make_local()}}
\item \href{#method-local_pc2-clone}{\code{local_pc2$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param2" data-id="draw_matrix"><a href='../../SimComm/html/community_param2.html#method-community_param2-draw_matrix'><code>SimComm::community_param2$draw_matrix()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param2" data-id="draw_wmppp"><a href='../../SimComm/html/community_param2.html#method-community_param2-draw_wmppp'><code>SimComm::community_param2$draw_wmppp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param2" data-id="set_values"><a href='../../SimComm/html/community_param2.html#method-community_param2-set_values'><code>SimComm::community_param2$set_values()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param2" data-id="show_values"><a href='../../SimComm/html/community_param2.html#method-community_param2-show_values'><code>SimComm::community_param2$show_values()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-local_pc2-new"></a>}}
\if{latex}{\out{\hypertarget{method-local_pc2-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_pc2$new(fashion = "matrix", death_rate = self$death_rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fashion}}{The fashion of the community (matrix or wmppp)}

\item{\code{death_rate}}{The mortality rate of an individual.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-local_pc2-make_local"></a>}}
\if{latex}{\out{\hypertarget{method-local_pc2-make_local}{}}}
\subsection{Method \code{make_local()}}{
Local community default community drawing - add death/birth_rate in the matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_pc2$make_local(
  nx = 10,
  ny = nx,
  S = 10,
  Distribution = "lnorm",
  sd = 1,
  prob = 0.1,
  theta = 40,
  death_rate = self$death_rate,
  fashion = "matrix"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{death_rate}}{The mortality rate of an individual.}

\item{\code{fashion}}{The fashion of the community (matrix or wmppp)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-local_pc2-clone"></a>}}
\if{latex}{\out{\hypertarget{method-local_pc2-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_pc2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
