% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Communities.R
\docType{class}
\name{meta_pc}
\alias{meta_pc}
\title{Meta community}
\description{
The class generate a meta community of a large number of species.
}
\section{Super class}{
\code{\link[SimComm:community_param]{SimComm::community_param}} -> \code{meta_pc}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{migration_rate}}{The migration of an individual.
Default is \code{0.005}}

\item{\code{speciation_rate}}{The speciation rate of an individual
Default is \code{0.00001}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-meta_pc-new}{\code{meta_pc$new()}}
\item \href{#method-meta_pc-make_meta}{\code{meta_pc$make_meta()}}
\item \href{#method-meta_pc-clone}{\code{meta_pc$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param" data-id="draw_matrix"><a href='../../SimComm/html/community_param.html#method-community_param-draw_matrix'><code>SimComm::community_param$draw_matrix()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param" data-id="draw_wmppp"><a href='../../SimComm/html/community_param.html#method-community_param-draw_wmppp'><code>SimComm::community_param$draw_wmppp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param" data-id="set_values"><a href='../../SimComm/html/community_param.html#method-community_param-set_values'><code>SimComm::community_param$set_values()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param" data-id="show_values"><a href='../../SimComm/html/community_param.html#method-community_param-show_values'><code>SimComm::community_param$show_values()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meta_pc-new"></a>}}
\if{latex}{\out{\hypertarget{method-meta_pc-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meta_pc$new(
  migration_rate = self$migration_rate,
  speciation_rate = self$speciation_rate
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{migration_rate}}{The migration rate of an individual.}

\item{\code{speciation_rate}}{The speciation rate of an individual.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meta_pc-make_meta"></a>}}
\if{latex}{\out{\hypertarget{method-meta_pc-make_meta}{}}}
\subsection{Method \code{make_meta()}}{
Meta community default community drawing - add migr/spec_rate in the matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meta_pc$make_meta(
  nx = 100,
  ny = nx,
  S = 100,
  Distribution = "lnorm",
  sd = 1,
  prob = 0.1,
  theta = 40,
  migration_rate = 0.05,
  speciation_rate = 0.001
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{migration_rate}}{The migration rate of an individual.}

\item{\code{speciation_rate}}{The speciation rate of an individual.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-meta_pc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-meta_pc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{meta_pc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
