% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{nnwhich.community_gridmodel}
\alias{nnwhich.community_gridmodel}
\title{Neighbor points in a grid model}
\usage{
nnwhich.community_gridmodel(X, neighborhood = "von Neumann")
}
\arguments{
\item{X}{A \code{\link{community_model}}.}

\item{neighborhood}{Definition of the neighborhood. "von Neumann" or "4" means the upper, lower, left and right neighbors are considered; "8" means the 8 nearest neighbors; "Moore" or "24" means the 24 nearest neighbors.}
}
\value{
A matrix. Each line corresponds to a point of the grid (points are ordered): values are the neighbor points.
}
\description{
Neighbor points in a grid model
}
