% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern.R
\name{pattern}
\alias{pattern}
\alias{pattern_grid}
\alias{pattern_matrix_individuals}
\alias{pattern_matrix_logical}
\alias{pm_Conway_blinker}
\alias{pm_Conway_glider}
\title{Pattern}
\usage{
pattern_grid(
  nx = 8,
  ny = nx,
  S = 300,
  Distribution = "lnorm",
  sd = 1,
  prob = 0.1,
  alpha = 40
)

pattern_matrix_individuals(
  nx = 8,
  ny = nx,
  S = 300,
  Distribution = "lnorm",
  sd = 1,
  prob = 0.1,
  alpha = 40
)

pattern_matrix_logical(nx = 8, ny = nx, prob = 0.5)

pm_Conway_blinker()

pm_Conway_glider(nx = 50, ny = nx)
}
\arguments{
\item{nx}{The number of X values (columns) of the grid or the matrix.}

\item{ny}{The number of Y values (lines) of the grid or the matrix.}

\item{S}{The number of species.}

\item{Distribution}{The distribution of species frequencies. May be "lnorm" (log-normal), "lseries" (log-series), "geom" (geometric) or "bstick" (broken stick).}

\item{sd}{The simulated distribution standard deviation. For the log-normal distribution, this is the standard deviation on the log scale.}

\item{prob}{The proportion of ressources taken by successive species in the geometric model.}

\item{alpha}{Fisher's alpha in the log-series model.}
}
\description{
The location of the agents of a model is described by a pattern.
}
\details{
Several patterns are available by default.
They may be point patterns (of class \link[spatstat.geom:ppp]{spatstat.geom::ppp}).
Regular point patterns are called grids.
They may also be matrices with no explicit coordinates by only notions of neighborhood.
Matrix models are by far faster to run.
Point patterns are provided by \link[SpatDiv:rSpCommunity]{SpatDiv::rSpCommunity}.
Patterns are to be used in the \code{pattern} field of a \link{community_model}.

Available patterns are:
\itemize{
\item \code{pattern_grid} A rectangular, regular grid of points.
Each point has marks.
Point are on a regular grid of 1x1 units of distance.
\item \code{pattern_matrix_individuals} A matrix.
Each cell of the matrix is an agent and has a numeric value equal to its type.
\item \code{pattern_matrix_logical} A matrix.
Each cell of the matrix is \code{FALSE} or \code{TRUE} (dead or alive, present or absent...).
\item \code{pm_Conway_blinker} A logical matrix for \href{https://en.wikipedia.org/wiki/Conway\\\%27s_Game_of_Life#Examples_of_patterns}{Conway's game of life}.
Three cells are alive.
The pattern is designed to oscillate.
\item \code{pm_Conway_glider} A logical matrix for \href{https://en.wikipedia.org/wiki/Conway\\\%27s_Game_of_Life#Examples_of_patterns}{Conway's game of life}.
The pattern is designed to move.
}
}
