% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{plot.community_model}
\alias{plot.community_model}
\title{plot Community Models}
\usage{
\method{plot}{community_model}(x, ...)
}
\arguments{
\item{x}{The model to plot.}

\item{...}{Extra arguments passed to the plot method.}
}
\description{
S3 method to force the use of the R6 method.
}
\details{
This is a helper function to always use the R6 method rather than the S3 method:
\code{plot(MyModel)} is translated to \code{MyModel$plot}.
}
\examples{
myModel <- cm_drift$new(pattern_matrix_individuals(S=10))
plot(myModel)
}
