% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropart.R
\name{Richness.pattern_matrix_individuals}
\alias{Richness.pattern_matrix_individuals}
\title{Richness of a matrix of individuals}
\usage{
\method{Richness}{pattern_matrix_individuals}(NorP, Correction = "Chao1",
  Alpha = 0.05, JackOver = FALSE, ..., CheckArguments = TRUE)
}
\arguments{
\item{NorP}{An object of class "wmppp" (\code{\link{pattern_matrix_individuals}}).}

\item{Correction}{A string containing one of the possible corrections: \code{"None"} (no correction), \code{"Jackknife"}, \code{"iChao1"}, or \code{"Chao1"}, the default value.}

\item{Alpha}{The risk level, 5\% by default, used to optimize the jackknife order.}

\item{JackOver}{If \code{TRUE}, retain the jackknife order immediately superior to the optimal one, usually resulting in the overestimation of the number of species. Default is \code{FALSE}.}

\item{...}{Further arguments. Unsused.}

\item{CheckArguments}{If \code{TRUE} (default), the function arguments are verified. Should be set to \code{FALSE} to save time in simulations for example, when the arguments have been checked elsewhere.}
}
\value{
A named number equal to the calculated diversity. The name is that of the bias correction used.
}
\description{
\code{Richness} is the number of species.
}
