% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropart.R
\name{Simpson.pattern_matrix_individuals}
\alias{Simpson.pattern_matrix_individuals}
\title{Simpson Entropy of a matrix of individuals}
\usage{
\method{Simpson}{pattern_matrix_individuals}(NorP, Correction = "Lande", ...,
  CheckArguments = TRUE)
}
\arguments{
\item{NorP}{An object of class "wmppp" (\code{\link{pattern_matrix_individuals}}).}

\item{Correction}{A string containing one of the possible corrections: see \code{\link{Simpson}}.}

\item{...}{Further arguments. Unsused.}

\item{CheckArguments}{If \code{TRUE} (default), the function arguments are verified. Should be set to \code{FALSE} to save time in simulations for example, when the arguments have been checked elsewhere.}
}
\value{
A named number equal to the calculated entropy. The name is that of the bias correction used.
}
\description{
Calculates the Simpson entropy of a \code{\link{pattern_matrix_individuals}} object.
}
