% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeLine.R
\name{timeline_regular}
\alias{timeline_regular}
\title{timeline_regular}
\usage{
timeline_regular(from = 1, to = 10, by = 1, unit = "Unit")
}
\arguments{
\item{from}{Time of the fist event}

\item{to}{Time of the last event}

\item{by}{Intervals between events}

\item{unit}{Unit of time}
}
\value{
a list of two items. \code{sequence} is a vector of numeric, positive values. \code{unit} is a string that describes the unit of time.
}
\description{
The moments when a \code{\link{community_model}} evolves.
}
\examples{
timeline_regular
}
