% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_Models.R
\docType{class}
\name{cm_drift}
\alias{cm_drift}
\title{Community Drift}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
cm_drift
}
\value{
An \code{\link{R6Class}}.
}
\description{
A \code{\link{community_model}} whose pattern is a regular, rectangular grid of points.
}
\section{Methods}{

\describe{
  \item{\code{initialize(pattern = pattern_matrix(), evolve = evolve_none, timeline = 0)}}{Initialization.}
  \item{\code{plot(...)}}{Default plot method: plots the pattern.}
  \item{\code{run(animate = FALSE, sleep = 0)}}{Run the model.}
}
}

\keyword{datasets}
