% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{community_matrixmodel}
\alias{community_matrixmodel}
\title{Community Matrix Model Class}
\value{
An \code{\link{R6Class}}.
}
\description{
A \code{\link{community_model}} whose pattern is a regular, rectangular grid of points.
}
\section{Public Methods}{

\describe{
  \item{\code{initialize(pattern = NULL, timeline = 0, type = "Species")}}{Initialization.}
  \item{\code{plot(time=NULL, sleep=animation::ani.options("interval"), ...)}}{Default plot method: plots the pattern.}
  \item{\code{autoplot(time=NULL, ...)}}{Makes a \code{\link{ggplot}} of the pattern.}
  \item{\code{run(animate = FALSE, sleep = animation::ani.options("interval"), save = FALSE, more_time = NULL)}}{Run the model.}
  \item{\code{saved_pattern(time)}}{Returns the pattern at the chosen time.}
  \item{\code{along_time(FUN, ...)}}{Applies the function FUN to the saved patterns along time and returns a dataframe with columns \code{x} for time and \code{y} for the results of FUN. FUN must return a single value.}
}
}

\examples{
myModel <- community_matrixmodel$new()
myModel$autoplot()
}
\section{Super class}{
\code{\link[SimComm:community_model]{SimComm::community_model}} -> \code{community_matrixmodel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-community_matrixmodel-new}{\code{community_matrixmodel$new()}}
\item \href{#method-community_matrixmodel-prepare_buffer}{\code{community_matrixmodel$prepare_buffer()}}
\item \href{#method-community_matrixmodel-neighbors}{\code{community_matrixmodel$neighbors()}}
\item \href{#method-community_matrixmodel-plot}{\code{community_matrixmodel$plot()}}
\item \href{#method-community_matrixmodel-autoplot}{\code{community_matrixmodel$autoplot()}}
\item \href{#method-community_matrixmodel-clone}{\code{community_matrixmodel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="along_time"><a href='../../SimComm/html/community_model.html#method-community_model-along_time'><code>SimComm::community_model$along_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="run"><a href='../../SimComm/html/community_model.html#method-community_model-run'><code>SimComm::community_model$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="saved_pattern"><a href='../../SimComm/html/community_model.html#method-community_model-saved_pattern'><code>SimComm::community_model$saved_pattern()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_matrixmodel-new"></a>}}
\if{latex}{\out{\hypertarget{method-community_matrixmodel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_matrixmodel$new(
  pattern = pattern_matrix_individuals(),
  timeline = 0,
  type = "Species",
  neighborhood = "von Neumann 1"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_matrixmodel-prepare_buffer"></a>}}
\if{latex}{\out{\hypertarget{method-community_matrixmodel-prepare_buffer}{}}}
\subsection{Method \code{prepare_buffer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_matrixmodel$prepare_buffer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_matrixmodel-neighbors"></a>}}
\if{latex}{\out{\hypertarget{method-community_matrixmodel-neighbors}{}}}
\subsection{Method \code{neighbors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_matrixmodel$neighbors(row, col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_matrixmodel-plot"></a>}}
\if{latex}{\out{\hypertarget{method-community_matrixmodel-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_matrixmodel$plot(time = NULL, sleep = 0, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_matrixmodel-autoplot"></a>}}
\if{latex}{\out{\hypertarget{method-community_matrixmodel-autoplot}{}}}
\subsection{Method \code{autoplot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_matrixmodel$autoplot(time = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_matrixmodel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-community_matrixmodel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_matrixmodel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
