% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{community_model}
\alias{community_model}
\title{Community Model Class}
\description{
The model represents a community, i.e. a set of interacting objects called agents.
Their location is described by a \code{pattern} that \code{evolve}s along a \code{timeline}.
}
\examples{
myModel <- community_matrixmodel$new()
myModel$autoplot()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pattern}}{The pattern which describes the location of agents.}

\item{\code{type}}{The type of individuals. Informational only.}

\item{\code{timeline}}{A numeric vector.}

\item{\code{last_time}}{The last time (in the time line) the model has been run.}

\item{\code{run_patterns}}{The past patterns of the model, obtained by \code{run} and saved.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-community_model-new}{\code{community_model$new()}}
\item \href{#method-community_model-plot}{\code{community_model$plot()}}
\item \href{#method-community_model-autoplot}{\code{community_model$autoplot()}}
\item \href{#method-community_model-run}{\code{community_model$run()}}
\item \href{#method-community_model-saved_pattern}{\code{community_model$saved_pattern()}}
\item \href{#method-community_model-along_time}{\code{community_model$along_time()}}
\item \href{#method-community_model-clone}{\code{community_model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-community_model-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_model$new(pattern = NULL, timeline = 0, type = "Species")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pattern}}{The pattern which describes the location of agents.}

\item{\code{timeline}}{A numeric vector that contains the points of time of interest.}

\item{\code{type}}{The type of individuals. Informational only.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_model-plot"></a>}}
\if{latex}{\out{\hypertarget{method-community_model-plot}{}}}
\subsection{Method \code{plot()}}{
Default plot method: plots the pattern.
To be overridden.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_model$plot(
  time = NULL,
  sleep = animation::ani.options("interval"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The point of the timeline considered.
Its value should be in \code{timeline}.}

\item{\code{sleep}}{The time (in ms) to sleep between each step of the evolution of the model.}

\item{\code{...}}{Extra arguments to be passed to methods.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_model-autoplot"></a>}}
\if{latex}{\out{\hypertarget{method-community_model-autoplot}{}}}
\subsection{Method \code{autoplot()}}{
Make a \link[ggplot2:ggplot]{ggplot2::ggplot} of the pattern.
To be overridden.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_model$autoplot(time = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The point of the timeline considered.
Its value should be in \code{timeline}.}

\item{\code{...}}{Extra arguments to be passed to methods.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_model-run"></a>}}
\if{latex}{\out{\hypertarget{method-community_model-run}{}}}
\subsection{Method \code{run()}}{
Run the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_model$run(
  animate = FALSE,
  sleep = animation::ani.options("interval"),
  save = FALSE,
  more_time = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{animate}}{if \code{TRUE}, the evolution of the model is shown in an animation.}

\item{\code{sleep}}{The time (in ms) to sleep between each step of the evolution of the model.}

\item{\code{save}}{if \code{TRUE}, the evolution of the model is saved.
if \code{FALSE}, only the final step is saved.}

\item{\code{more_time}}{A numeric vector that extends the timeline.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_model-saved_pattern"></a>}}
\if{latex}{\out{\hypertarget{method-community_model-saved_pattern}{}}}
\subsection{Method \code{saved_pattern()}}{
Return the pattern at the chosen time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_model$saved_pattern(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The point of the timeline considered.
Its value should be in \code{timeline}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_model-along_time"></a>}}
\if{latex}{\out{\hypertarget{method-community_model-along_time}{}}}
\subsection{Method \code{along_time()}}{
Apply the function \code{FUN} to the saved patterns along time and return
a dataframe with columns \code{x} for time and \code{y} for the results of \code{FUN}.
\code{FUN} must return a single value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_model$along_time(FUN, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FUN}}{A function to apply to each pattern of the evolution of the model.
It must return a single, numeric value.}

\item{\code{...}}{Extra arguments to be passed to methods.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-community_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
