% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{community_model}
\alias{community_model}
\title{Community Model Class}
\format{\code{\link{R6Class}} object.}
\usage{
community_model
}
\value{
An \code{\link{R6Class}}.
}
\description{
The model represents a community, i.e. a set of interacting objects called agents. Their location is described by a \code{pattern} that \code{evolve}s along a \code{timeline}.
}
\section{Fields}{

\describe{
\item{\code{pattern}}{The pattern which describes the location of agents.}

\item{\code{type}}{The type of individuals. Informational only.}

\item{\code{timeline}}{A numeric vector.}

\item{\code{last_time}}{The last time (in the time line) the model has been run.}

\item{\code{run_patterns}}{The past patterns of the model, obtained by \code{run} and saved.}

\item{\code{neighborhood}}{A character string defining what is the neighborhood of a cell:
"von Neumann 1" or "4" for the closest four neighbors (North, West, South, East);
"Moore 1" or "8" for all adjacent cells (the first four and North-West, etc.);
"Moore 2" or "24" for two rings of neighbors.}
}}

\section{Public Methods}{

\describe{
  \item{\code{initialize(pattern = NULL, timeline = 0, type = "Species")}}{Initialization.}
  \item{\code{plot(time=NULL, sleep=animation::ani.options("interval"), ...)}}{Default plot method: plots the pattern.}
  \item{\code{autoplot(time=NULL, ...)}}{Makes a \code{\link{ggplot}} of the pattern.}
  \item{\code{run(animate = FALSE, sleep = animation::ani.options("interval"), save = FALSE, more_time = NULL)}}{Run the model.}
  \item{\code{saved_pattern(time)}}{Returns the pattern at the chosen time.}
  \item{\code{along_time(FUN, ...)}}{Applies the function FUN to the saved patterns along time and returns a dataframe with columns \code{x} for time and \code{y} for the results of FUN. FUN must return a single value.}
}
}

\examples{
myModel <- community_matrixmodel$new()
myModel$autoplot()
}
\keyword{datasets}
