% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{community_model}
\alias{community_model}
\title{Community Model Class}
\format{\code{\link{R6Class}} object.}
\usage{
community_model
}
\value{
An \code{\link{R6Class}}.
}
\description{
The model represents a community, i.e. a set of interacting objects called agents. Their location is described by a \code{pattern} that \code{evolve}s along a \code{timeline}.
}
\section{Fields}{

\describe{
\item{\code{pattern}}{The pattern which describes the location of agents.}

\item{\code{timeline}}{A numeric vector.}
}}

\section{Methods}{

\describe{
  \item{\code{initialize(pattern = pattern_matrix(), evolve = evolve_none, timeline = 0)}}{Initialization.}
  \item{\code{plot(...)}}{Default plot method: plots the pattern.}
  \item{\code{run(animate = FALSE, sleep = 0)}}{Run the model.}
}
}

\examples{
myModel <- community_model$new
}
\keyword{datasets}
