% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Communities.R
\docType{class}
\name{community_param}
\alias{community_param}
\title{Community Attributes Class}
\description{
The class generate a community of a large number of species.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-community_param-new}{\code{community_param$new()}}
\item \href{#method-community_param-set_values}{\code{community_param$set_values()}}
\item \href{#method-community_param-show_values}{\code{community_param$show_values()}}
\item \href{#method-community_param-draw_matrix}{\code{community_param$draw_matrix()}}
\item \href{#method-community_param-draw_wmppp}{\code{community_param$draw_wmppp()}}
\item \href{#method-community_param-clone}{\code{community_param$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_param-new"></a>}}
\if{latex}{\out{\hypertarget{method-community_param-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_param$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_param-set_values"></a>}}
\if{latex}{\out{\hypertarget{method-community_param-set_values}{}}}
\subsection{Method \code{set_values()}}{
Set values of the community generation parameters (set to default if no arguments)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_param$set_values(
  nx = 20,
  ny = nx,
  S = 20,
  Distribution = "lnorm",
  sd = 1,
  prob = 0.1,
  alpha = 40,
  distribute = "entropart",
  style = "random"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nx}}{The size of horizontal axis}

\item{\code{ny}}{The size of vertical axis}

\item{\code{S}}{The number of species.}

\item{\code{Distribution}}{The distribution of species frequencies. May be "lnorm" (log-normal), "lseries" (log-series), "geom" (geometric) or "bstick" (broken stick).}

\item{\code{sd}}{The simulated distribution standard deviation. For the log-normal distribution, this is the standard deviation on the log scale.}

\item{\code{prob}}{The proportion of ressources taken by successive species in the geometric model.}

\item{\code{alpha}}{Fisher's alpha in the log-series model.}

\item{\code{distribute}}{The type of distribution. May be "entropart" (entropart package), "random" or "uniform".}

\item{\code{style}}{The style of the community. May be "random", "checkerboard" or "uniform".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_param-show_values"></a>}}
\if{latex}{\out{\hypertarget{method-community_param-show_values}{}}}
\subsection{Method \code{show_values()}}{
Show the values of the community generation parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_param$show_values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_param-draw_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-community_param-draw_matrix}{}}}
\subsection{Method \code{draw_matrix()}}{
Draw a community
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_param$draw_matrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_param-draw_wmppp"></a>}}
\if{latex}{\out{\hypertarget{method-community_param-draw_wmppp}{}}}
\subsection{Method \code{draw_wmppp()}}{
Draw a wmppp community with no adjective
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_param$draw_wmppp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_param-clone"></a>}}
\if{latex}{\out{\hypertarget{method-community_param-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_param$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
