% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{community_spcmodel}
\alias{community_spcmodel}
\title{Community Point Pattern Class}
\description{
Community Point Pattern Class

Community Point Pattern Class
}
\details{
A \link{community_model} whose pattern is a \link[dbmss:wmppp]{dbmss::wmppp} object..
}
\section{Super class}{
\code{\link[SimComm:community_model]{SimComm::community_model}} -> \code{community_spcmodel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tess}}{The Dirichlet tessellation of the spatial point pattern.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-community_spcmodel-new}{\code{community_spcmodel$new()}}
\item \href{#method-community_spcmodel-neighbors_n}{\code{community_spcmodel$neighbors_n()}}
\item \href{#method-community_spcmodel-neighbors_r}{\code{community_spcmodel$neighbors_r()}}
\item \href{#method-community_spcmodel-neighbor_types_r}{\code{community_spcmodel$neighbor_types_r()}}
\item \href{#method-community_spcmodel-plot}{\code{community_spcmodel$plot()}}
\item \href{#method-community_spcmodel-clone}{\code{community_spcmodel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="along_time"><a href='../../SimComm/html/community_model.html#method-community_model-along_time'><code>SimComm::community_model$along_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="autoplot"><a href='../../SimComm/html/community_model.html#method-community_model-autoplot'><code>SimComm::community_model$autoplot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="run"><a href='../../SimComm/html/community_model.html#method-community_model-run'><code>SimComm::community_model$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="saved_pattern"><a href='../../SimComm/html/community_model.html#method-community_model-saved_pattern'><code>SimComm::community_model$saved_pattern()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_spcmodel-new"></a>}}
\if{latex}{\out{\hypertarget{method-community_spcmodel-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$new(
  pattern = SpatDiv::rSpCommunity(n = 1, size = 100, CheckArguments = FALSE),
  timeline = 0,
  type = "Species"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pattern}}{The pattern which describes the location of agents.}

\item{\code{timeline}}{A numeric vector that contains the points of time of interest.}

\item{\code{type}}{The type of individuals. Informational only.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_spcmodel-neighbors_n"></a>}}
\if{latex}{\out{\hypertarget{method-community_spcmodel-neighbors_n}{}}}
\subsection{Method \code{neighbors_n()}}{
The n nearest neighbors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$neighbors_n(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{The number of neighbors considered as in the neighborhood.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_spcmodel-neighbors_r"></a>}}
\if{latex}{\out{\hypertarget{method-community_spcmodel-neighbors_r}{}}}
\subsection{Method \code{neighbors_r()}}{
Neighbors less than r apart
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$neighbors_r(r, keep_distances = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{The distance that defines the neighborhood.}

\item{\code{keep_distances}}{if \code{TRUE}, save the vector of distances to neighbors in the attribute \code{the_neighbors}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_spcmodel-neighbor_types_r"></a>}}
\if{latex}{\out{\hypertarget{method-community_spcmodel-neighbor_types_r}{}}}
\subsection{Method \code{neighbor_types_r()}}{
The type of neighbors less than r apart
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$neighbor_types_r(r)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{The distance that defines the neighborhood.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_spcmodel-plot"></a>}}
\if{latex}{\out{\hypertarget{method-community_spcmodel-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the pattern.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$plot(time = NULL, sleep = 0, which.marks = "PointType", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The point of the timeline considered.}

\item{\code{sleep}}{The time (in ms) to sleep between each step of the evolution of the model.}

\item{\code{which.marks}}{The marks to plot, that must be a the name of a column of the mark dataframe.}

\item{\code{...}}{Extra arguments to be passed to methods.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_spcmodel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-community_spcmodel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
