% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{community_spcmodel}
\alias{community_spcmodel}
\title{Community Point Pattern Class}
\value{
An \code{\link{R6Class}}.
}
\description{
A \code{\link{community_model}} whose pattern is a \code{\link[dbmss]{wmppp}} object..
}
\section{Public Methods}{

\describe{
  \item{\code{initialize(pattern = NULL, timeline = 0, type = "Species")}}{Initialization.}
  \item{\code{plot(time=NULL, sleep=animation::ani.options("interval"), ...)}}{Default plot method: plots the pattern.}
  \item{\code{autoplot(time=NULL, ...)}}{Makes a \code{\link{ggplot}} of the pattern.}
  \item{\code{run(animate = FALSE, sleep = animation::ani.options("interval"), save = FALSE, more_time = NULL)}}{Run the model.}
  \item{\code{saved_pattern(time)}}{Returns the pattern at the chosen time.}
  \item{\code{along_time(FUN, ...)}}{Applies the function FUN to the saved patterns along time and returns a dataframe with columns \code{x} for time and \code{y} for the results of FUN. FUN must return a single value.}
}
}

\examples{
myModel <- community_matrixmodel$new()
myModel$autoplot()
}
\section{Super class}{
\code{\link[SimComm:community_model]{SimComm::community_model}} -> \code{community_spcmodel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-community_spcmodel-new}{\code{community_spcmodel$new()}}
\item \href{#method-community_spcmodel-neighbors_n}{\code{community_spcmodel$neighbors_n()}}
\item \href{#method-community_spcmodel-neighbors_r}{\code{community_spcmodel$neighbors_r()}}
\item \href{#method-community_spcmodel-neighbor_types_r}{\code{community_spcmodel$neighbor_types_r()}}
\item \href{#method-community_spcmodel-plot}{\code{community_spcmodel$plot()}}
\item \href{#method-community_spcmodel-clone}{\code{community_spcmodel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="along_time"><a href='../../SimComm/html/community_model.html#method-community_model-along_time'><code>SimComm::community_model$along_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="autoplot"><a href='../../SimComm/html/community_model.html#method-community_model-autoplot'><code>SimComm::community_model$autoplot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="run"><a href='../../SimComm/html/community_model.html#method-community_model-run'><code>SimComm::community_model$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="saved_pattern"><a href='../../SimComm/html/community_model.html#method-community_model-saved_pattern'><code>SimComm::community_model$saved_pattern()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_spcmodel-new"></a>}}
\if{latex}{\out{\hypertarget{method-community_spcmodel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$new(
  pattern = SpatDiv::rSpCommunity(n = 1, size = 100, CheckArguments = FALSE),
  timeline = 0,
  type = "Species"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_spcmodel-neighbors_n"></a>}}
\if{latex}{\out{\hypertarget{method-community_spcmodel-neighbors_n}{}}}
\subsection{Method \code{neighbors_n()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$neighbors_n(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_spcmodel-neighbors_r"></a>}}
\if{latex}{\out{\hypertarget{method-community_spcmodel-neighbors_r}{}}}
\subsection{Method \code{neighbors_r()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$neighbors_r(r, keep_distances = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_spcmodel-neighbor_types_r"></a>}}
\if{latex}{\out{\hypertarget{method-community_spcmodel-neighbor_types_r}{}}}
\subsection{Method \code{neighbor_types_r()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$neighbor_types_r(r)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_spcmodel-plot"></a>}}
\if{latex}{\out{\hypertarget{method-community_spcmodel-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$plot(time = NULL, sleep = 0, which.marks = "PointType", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-community_spcmodel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-community_spcmodel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{community_spcmodel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
