% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_Models.R
\name{cp_drift}
\alias{cp_drift}
\title{Community Drift}
\description{
Community Drift

Community Drift
}
\details{
A \link{community_spcmodel}.
At each generation, each individual is replaced by one of its neighbors.
}
\section{Super classes}{
\code{\link[SimComm:community_model]{SimComm::community_model}} -> \code{\link[SimComm:community_spcmodel]{SimComm::community_spcmodel}} -> \code{cp_drift}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n_neighbors}}{The number of nearest neighbors to take into account.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cp_drift-new}{\code{cp_drift$new()}}
\item \href{#method-cp_drift-clone}{\code{cp_drift$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="along_time"><a href='../../SimComm/html/community_model.html#method-community_model-along_time'><code>SimComm::community_model$along_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="autoplot"><a href='../../SimComm/html/community_model.html#method-community_model-autoplot'><code>SimComm::community_model$autoplot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="run"><a href='../../SimComm/html/community_model.html#method-community_model-run'><code>SimComm::community_model$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_model" data-id="saved_pattern"><a href='../../SimComm/html/community_model.html#method-community_model-saved_pattern'><code>SimComm::community_model$saved_pattern()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_spcmodel" data-id="neighbor_types_r"><a href='../../SimComm/html/community_spcmodel.html#method-community_spcmodel-neighbor_types_r'><code>SimComm::community_spcmodel$neighbor_types_r()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_spcmodel" data-id="neighbors_n"><a href='../../SimComm/html/community_spcmodel.html#method-community_spcmodel-neighbors_n'><code>SimComm::community_spcmodel$neighbors_n()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_spcmodel" data-id="neighbors_r"><a href='../../SimComm/html/community_spcmodel.html#method-community_spcmodel-neighbors_r'><code>SimComm::community_spcmodel$neighbors_r()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_spcmodel" data-id="plot"><a href='../../SimComm/html/community_spcmodel.html#method-community_spcmodel-plot'><code>SimComm::community_spcmodel$plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cp_drift-new"></a>}}
\if{latex}{\out{\hypertarget{method-cp_drift-new}{}}}
\subsection{Method \code{new()}}{
Create a new object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cp_drift$new(
  pattern = SpatDiv::rSpCommunity(n = 1, size = 100, CheckArguments = FALSE),
  timeline = 0,
  type = "Species",
  n_neighbors = 6
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pattern}}{The pattern which describes the location of agents.}

\item{\code{timeline}}{A numeric vector that contains the points of time of interest.}

\item{\code{type}}{The type of individuals. Informational only.}

\item{\code{n_neighbors}}{Number of neighbors.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cp_drift-clone"></a>}}
\if{latex}{\out{\hypertarget{method-cp_drift-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cp_drift$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
