% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Communities.R
\docType{class}
\name{local_pc}
\alias{local_pc}
\title{Local community}
\description{
The class generate a local community of a large number of species.
}
\section{Super class}{
\code{\link[SimComm:community_param]{SimComm::community_param}} -> \code{local_pc}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{death_rate}}{The mortality rate of an individual.
Default is \code{0.1}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-local_pc-new}{\code{local_pc$new()}}
\item \href{#method-local_pc-make_local}{\code{local_pc$make_local()}}
\item \href{#method-local_pc-clone}{\code{local_pc$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param" data-id="draw_matrix"><a href='../../SimComm/html/community_param.html#method-community_param-draw_matrix'><code>SimComm::community_param$draw_matrix()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param" data-id="draw_wmppp"><a href='../../SimComm/html/community_param.html#method-community_param-draw_wmppp'><code>SimComm::community_param$draw_wmppp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param" data-id="set_values"><a href='../../SimComm/html/community_param.html#method-community_param-set_values'><code>SimComm::community_param$set_values()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SimComm" data-topic="community_param" data-id="show_values"><a href='../../SimComm/html/community_param.html#method-community_param-show_values'><code>SimComm::community_param$show_values()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-local_pc-new"></a>}}
\if{latex}{\out{\hypertarget{method-local_pc-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_pc$new(fashion = "matrix", death_rate = self$death_rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fashion}}{The fashion of the community (matrix or wmppp)}

\item{\code{death_rate}}{The mortality rate of an individual.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-local_pc-make_local"></a>}}
\if{latex}{\out{\hypertarget{method-local_pc-make_local}{}}}
\subsection{Method \code{make_local()}}{
Local community default community drawing - add death/birth_rate in the matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_pc$make_local(
  nx = 10,
  ny = nx,
  S = 10,
  Distribution = "lnorm",
  sd = 1,
  prob = 0.1,
  theta = 40,
  death_rate = self$death_rate,
  fashion = "matrix"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{death_rate}}{The mortality rate of an individual.}

\item{\code{fashion}}{The fashion of the community (matrix or wmppp)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-local_pc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-local_pc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{local_pc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
