% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern.R
\name{pattern}
\alias{pattern}
\alias{pattern_grid}
\alias{pattern_matrix}
\title{Pattern}
\usage{
pattern_grid(nx = 8, ny = nx)

pattern_matrix(nx = 8, ny = nx)
}
\arguments{
\item{nx}{The number of X values (columns) of the grid or the matrix.}

\item{ny}{The number of y values (lines) of the grid or the matrix.}
}
\value{
\code{evolve} methods should modifiy the model's pattern directly and return \code{NULL}.
}
\description{
The location of the agents of a model is described by a pattern.
Several patterns are available by default.
They may be point patterns (of class \code{\link{ppp}}). Regular point patterns are called grids.
They may also be matrices with no explicit coordinates by only notions of neighborhood. Matrix models are by far faster to run.
Patterns are to be used in the \code{pattern} field of a \code{\link{community_model}}.
}
\details{
Available patterns:
\describe{
  \item{\code{pattern_grid}}{A rectangular, regular grid of points. Each point has marks. Point are on a regular grid of 1x1 units of distance.}
  \item{\code{pattern_matrix}}{A matrix. Each cell of the matrix has a numeric value.}
}
}
